split files by class
